Improve default behavior of sp_insertable, sp_prob
Prior to this commit, sp_insertable and sp_prob_swap would
be set to True/1.0 even in the case of NVT or NpT simulations.
This is possibly confusing since no insertion/swap moves are
attempted in those cases. Now, sp_insertable is set to False
and sp_prob_swap is set to 0.0 unless in an open ensemble.