Fix box size modified if run is within a loop
The box size was modified by a factor of 10
if the run function was within a loop.

The variable box_matrix was being assigned to
the vectors attribute of the system object.
Then, box_matrix (which is also an object
reference to box.box.vectors) was being converted
to nanometers, therefore introducing a factor
of 10 each time an iteration of the loop
completed.